feat: show and search for space attributes in Booking UI
